TSM-53 - web controller logging is way too verbose! * fixed unit test.  It works inside myeclipse but gets a compile error from the build file.  argh.

+++ trunk/tsm/src/test/java/com/tech4d/tsm/service/IntegrationTestSearchBoundaries.java
@@ -116,9 +116,9 @@ public class IntegrationTestSearchBoundaries {
     public void testGetCount() throws ParseException, java.text.ParseException {
         makeSearchEvent(insideTheBox, TimeRangeFormat.parse("1522-1527"), "Stuff", null);
         makeSearchEvent(insideTheBox, TimeRangeFormat.parse("1522-1527"), "Stuff", null);
-        assertEquals(2, eventSearchService.getTotalCount());
+        assertEquals((Integer)2, eventSearchService.getTotalCount());
         makeSearchEvent(insideTheBox, TimeRangeFormat.parse("1522-1527"), "Stuff", null);
-        assertEquals(3, eventSearchService.getTotalCount());
+        assertEquals((Integer)3, eventSearchService.getTotalCount());
     	
     }
